Move internal data structures from PoolChunk to top-level classes
Motivation:
We wish to test these independently.

Modification:
Make inner classes top-level, moving LongLongHashMap and LongPriorityQueue out of PoolChunk.

Result:
These classes can now be accessed from tests.
