+++ buffer/src/main/java/io/netty/buffer/PoolChunk.java
@@ -17,7 +17,6 @@ package io.netty.buffer;
 
 import java.nio.ByteBuffer;
 import java.util.ArrayDeque;
-import java.util.Arrays;
 import java.util.Deque;
 import java.util.PriorityQueue;
 
@@ -637,162 +636,5 @@ final class PoolChunk<T> implements PoolChunkMetric {
 
     static int bitmapIdx(long handle) {
         return (int) handle;
-    }
-
-    private static final class LongLongHashMap {
-        private static final int MAX_PROBE = 4;
-        private int mask = 31;
-        private long[] array = new long[mask + 1];
-        private long zeroVal;
-
-        public long put(long key, long value) {
-            if (key == 0) {
-                long prev = zeroVal;
-                zeroVal = value;
-                return prev;
-            }
-
-            for (;;) {
-                int index = index(key);
-                for (int i = 0; i < MAX_PROBE; i++) {
-                    long existing = array[index];
-                    if (existing == key || existing == 0) {
-                        long prev = array[index + 1];
-                        array[index] = key;
-                        array[index + 1] = value;
-                        return prev;
-                    }
-                    index = index + 2 & mask;
-                }
-                expand(); // Grow array and re-hash.
-            }
-        }
-
-        public void remove(long key) {
-            if (key == 0) {
-                zeroVal = 0;
-                return;
-            }
-            int index = index(key);
-            for (int i = 0; i < MAX_PROBE; i++) {
-                long existing = array[index];
-                if (existing == key) {
-                    array[index] = 0;
-                    array[index + 1] = 0;
-                    break;
-                }
-                index = index + 2 & mask;
-            }
-        }
-
-        public long get(long key) {
-            if (key == 0) {
-                return zeroVal;
-            }
-            int index = index(key);
-            for (int i = 0; i < MAX_PROBE; i++) {
-                long existing = array[index];
-                if (existing == key) {
-                    return array[index + 1];
-                }
-                index = index + 2 & mask;
-            }
-            return -1;
-        }
-
-        private int index(long key) {
-            return (int) (key << 1 & mask);
-        }
-
-        private void expand() {
-            long[] prev = array;
-            int newSize = prev.length * 2;
-            mask = newSize - 1;
-            array = new long[newSize];
-            for (int i = 0; i < prev.length >> 1; i += 2) {
-                long key = prev[i];
-                long val = prev[i + 1];
-                put(key, val);
-            }
-        }
-    }
-
-    private static final class LongPriorityQueue {
-        private long[] array = new long[9];
-        private int size;
-
-        public void offer(long handle) {
-            size++;
-            if (size == array.length) {
-                // Grow queue capacity.
-                array = Arrays.copyOf(array, 1 + (array.length - 1) * 2);
-            }
-            array[size] = handle;
-            lift();
-        }
-
-        public void remove(long value) {
-            for (int i = 1; i <= size; i++) {
-                if (array[i] == value) {
-                    if (i == size) {
-                        array[i] = 0;
-                    } else {
-                        array[i] = array[size];
-                        sink(i);
-                    }
-                    size--;
-                    return;
-                }
-            }
-        }
-
-        public long poll() {
-            if (size == 0) {
-                return 0;
-            }
-            long val = array[1];
-            array[1] = array[size];
-            array[size] = 0;
-            size--;
-            sink(1);
-            return val;
-        }
-
-        public boolean isEmpty() {
-            return size == 0;
-        }
-
-        private void lift() {
-            int index = size;
-            int parentIndex;
-            while (index > 1 && subord(parentIndex = index >> 1, index)) {
-                swap(index, parentIndex);
-                index = parentIndex;
-            }
-        }
-
-        private void sink(int index) {
-            int child;
-            while ((child = index << 1) <= size) {
-                if (child < size && subord(child, child + 1)) {
-                    child++;
-                }
-                if (!subord(index, child)) {
-                    break;
-                }
-                swap(index, child);
-                index = child;
-            }
-        }
-
-        private boolean subord(int a, int b) {
-            return array[a] > array[b];
-        }
-
-        private void swap(int a, int b) {
-            long value = array[a];
-            array[a] = array[b];
-            array[b] = value;
-        }
     }
 }
